Refactor Unit Tests for OpenCL Kernels
This pull request refactors the unit tests for OpenCL kernels.
The changes involve splitting the test files into three categories based on functionality.
One for INT4-related tests, another for QK_K (GGML format) tests, and a third for the remaining BLAS tests.
Additionally, it consolidates common helper functions used across the test files into separate utility header files.

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped
